requirements.txt
bottle==0.12.18
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==3.2
dataclasses==0.6
decorator==4.4.2
future==0.18.2
idna==2.9
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.2
pycparser==2.20
pyparsing==2.4.7
pyTelegramBotAPI==3.6.7
python-dateutil==2.8.1
python-telegram-bot==12.6.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
gunicorn==19.9.0
